+++ Imager/imager.client/src/app/app.module.ts
@@ -4,7 +4,7 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { UploadFormComponent } from './upload-form/upload-form.component';
+import { UploadFormComponent } from './features/upload/upload-form/upload-form.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
 
@@ -14,7 +14,7 @@ import { MatButtonModule } from '@angular/material/button'
 import { MatCardModule } from '@angular/material/card'
 import { MatIconModule } from '@angular/material/icon'
 import { MatDividerModule } from '@angular/material/divider'
-
+import { MatProgressBar, MatProgressBarModule } from '@angular/material/progress-bar'
 
 @NgModule({
   declarations: [
@@ -23,16 +23,14 @@ import { MatDividerModule } from '@angular/material/divider'
   ],
   imports: [
     BrowserModule,
-    FormsModule,
     ReactiveFormsModule,
-
     MatFormFieldModule,
     MatInputModule,
     MatButtonModule,
     MatCardModule,
     MatIconModule,
     MatDividerModule,
-
+    MatProgressBarModule,
     HttpClientModule,
     AppRoutingModule
   ],