+++ Imager/Imager.Server/CHANGELOG.md
@@ -6,3 +6,6 @@ The following steps were used to generate this project:
 - Update project file to add a reference to the frontend project and set SPA properties.
 - Add project to the startup projects list.
 - Write this file.
+
+
+- Remove `WeatherForecast.cs` file and anything related to Forecasting